Move MarkQueue to its own file
