+++ packages/dds/tree/src/feature-libraries/sequence-field/compose.ts
@@ -17,10 +17,10 @@ import {
     SizedObjectMark,
 } from "./format";
 import { MarkListFactory } from "./markListFactory";
+import { MarkQueue } from "./markQueue";
 import {
     applyMoveEffectsToMark,
     changeSrcMoveId,
-    MarkQueue,
     modifyMoveSrc,
     MoveEffectTable,
     newMoveEffectTable,
